% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanLattice-function.R, R/create-latex-tablefunction.R
\name{clanLattice}
\alias{clanLattice}
\alias{create.latex.table}
\title{Version function for the clanLattice library}
\usage{
clanLattice()

create.latex.table()
}
\value{
A list of information about the version and functions within clanLattice.

create.latex.table.
}
\description{
Version function for the clanLattice library.

create.latex.table.
}
\author{
Claus E. Andersen

Claus E. Andersen
}

