% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-latex-tablefunction.R
\name{create.latex.table}
\alias{create.latex.table}
\title{Title create.latex.table}
\usage{
create.latex.table(df, pretty.func = function(x, col.num = 1) {     x },
  NL = "\\\\\\\\", col.sign = "&")
}
\arguments{
\item{df}{}

\item{pretty.func}{= function(x, col.num = 1){x}}

\item{NL}{}

\item{col.sign}{}
}
\description{
Title create.latex.table
}
\examples{
whatever
}

