% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left-aligned-main-label-demo-function.R
\name{left.aligned.main.label.demo}
\alias{left.aligned.main.label.demo}
\title{Demonstration how tocreate a lattice plot with a left-aligned main label}
\usage{
require(lattice)
require(grid)
plt <- left.aligned.main.label.demo()
print(plt)
}
\value{
A Lattice plot
}
\description{
This functions creates a left-aligned main label
using a textgrob as given below.
How did I get that idea? See Lattice book p.152 / sec. 9.2.1 + demo("labels",package="lattice")
}
\author{
Claus E. Andersen
}
