% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-binned-errorbars-demo-function.R
\name{panel.binned.errorbars.demo}
\alias{panel.binned.errorbars.demo}
\title{Demonstration of the use of panel.binned.errorbars for Lattice graphics}
\usage{
panel.binned.errorbars.demo()
}
\arguments{
\item{dat}{is not used for anything}

\item{dat.ok}{is not used for anything}

\item{return.synthetic.data}{is not used for anything}
}
\value{
The main output is the sideeffect of creating a bunch of lattice plots
}
\description{
Demonstration of the use of panel.binned.errorbars for Lattice graphics.
}
\author{
Claus E. Andersen
}

