% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-binned-errorbars-function.R
\name{panel.binned.errorbars}
\alias{panel.binned.errorbars}
\title{Lattice plotting panel for binned data (e.g. aggregared standard deviations)}
\usage{
see the body of the function plut the function called panel.binned.errorbars.demo
}
\arguments{
\item{x}{is the x-coordinates for the data}

\item{y}{is the y-coordinates od the data}
}
\value{
plt A Lattice plot showing the colors
}
\description{
This is a panel function for Lattice plotting. The data is binned in accordance with the 
x-values and calculate standard deviations, for example.
}
\author{
Claus E. Andersen
}
