% Generated by roxygen2 (4.0.1): do not edit by hand
\name{panel.ebars.demo}
\alias{panel.ebars.demo}
\title{Demonstration of how to use panel.ebars for drawing error bars in lattice plots}
\usage{
panel.ebars.demo()
}
\value{
Lattice plot examples with error bars
}
\description{
Demonstration of how to use panel.ebars for drawing error bars in lattice plots
}
\author{
Claus E. Andersen
}

