% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-ebars-function.R
\name{panel.ebars.grouped}
\alias{panel.ebars.grouped}
\title{Compute groupwise standard deviations and plot errorbars (lattice panel)}
\usage{
See above
}
\arguments{
\item{x}{= x-coordinate for the data points.}

\item{y}{= y-coordinate for the data points.}

\item{type}{(only one option is available, "sd").}

\item{err.col}{is the color of the error bar (and mean symbol)}

\item{err.type}{is the type of plot for the mean values (e.g. "n", "p", or "b").}

\item{err.pch}{is the plotting symbol for mean values.}

\item{err.cex}{is the size of the plotting symbol for the mean values.}

\item{err.lwd}{is the line width for the error bars.}

\item{err.width}{is the width of the error bars (i.e. the wiskers) in mm.}

\item{err.offsetis}{an offset for the error bar relative to the mean (x).}
}
\value{
A Lattice panel
}
\description{
This function has been designed for plotting of errorbars in lattice plots.
This is an alternative to the more complex panel-function called \link{panel.binned.errorbars}.
Example:

  N <- 50
  
  df <- data.frame(degC = sample(20:25,N,replace=TRUE),counts=NA,instrument=sample(c("Fluke","HP"),N,replace=TRUE))
  
  df$counts <- df$degC * 1.1 + rnorm(nrow(df),mean=0,sd=0.3)
  
  
  xyplot(counts ~ degC|instrument,
    data=df,
    panel=function(x,y,...)\{
    panel.xyplot(x,y,...); 
    panel.ebars.grouped(x,y,type="sd",err.type="b",err.col="red",err.pch=16,err.cex=0.8,err.lwd=1,err.width=1.2,err.offset=0) 
    \})
    
See also \link{panel.binned.errorbars}, \link{panel.binned.errorbars}, \link{panel.ebars} and \link{panel.ebars2}.
}
\author{
Claus E. Andersen
}
