% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-ebars-function.R
\name{panel.ebars2}
\alias{panel.ebars2}
\title{Improved errorbar panel for Lattice graphics (version 2)}
\usage{
require(lattice)
require(clanLattice)
require (grid)
set.trellis(pch=16)
plt <- dotplot(Species ~ Sepal.Length,
data=iris,
main="Simple demonstration of the panel.ebars2 function"
groups=Species,
auto.key=list(columns=3),
y.err =iris$Sepal.Width/10, 
panel = function(x, y, subscripts, groups, ...) {
  yy <- as.numeric(y)
  panel.ebars2(x, yy, subscripts, 
     y.wanted = !FALSE, 
     x.wanted = FALSE, 
     point.wanted = !FALSE,...)
  panel.superpose(x, y, subscripts, groups, ...)
  }
  ) # dotplot
  print(plt)
# See the function panel.ebars.demo for further details
}
\arguments{
\item{x}{= x-coordinate for the data points.}

\item{y}{= y-coordinate for the data points.}

\item{x.wanted}{= TRUE means that the x-errorbars will be plotted.}

\item{y.wanted}{= TRUE means that the y-errorbars will be plotted.}

\item{x.err}{= the size of symmetrical x-errorbars (i.e. x.err = x.minus = x.plus).
Note, that x.err, x.minus, or x.plus must be of the same length as 
x and y. This is so because these vectors are manipulated deeper down 
using the subscripts argument. Also, note that it is not enough to just
set x.err etc. equal to a named column in the main dataframe. So, if the
main dataframe is called df, and if it contains a column called u.Gy that 
you want to use as errorbar, then you must write x.err = df$u.Gy (it is
NOT enough to write x.err = u.Gy even though data=df). Likewise, if you
want to set the errorbar to a constant value, you must use
something like x.err = rep(2.3,nrow(df)) rather than just x.err = 2.3.}

\item{x.minus}{= the size of the lower x-errorbars (if assymetrical)}

\item{x.plus}{= the size of the upper x-errorbars (if assymetrical)}

\item{x.width}{= the width (in mm) of the whiskers on the x-errorbars.}

\item{x.inner}{= the free space (in mm) around the main data point where no part of the errorbas should be drawn.}

\item{x.offset}{= the offset (in mm) of the x-errorbar.}

\item{gp.x}{= graphical parameters for the x-errorbars.}

\item{gp.xwhisker}{= graphical parameters for the whiskers for the x-errorbars.}

\item{point.wanted}{= TRUE means that a separate point will be drawn at the central data.point.}

\item{gp.point}{= graphical parameters for the separate data point to be drawn.}
}
\value{
A Lattice panel
}
\description{
This function has been designed for plotting of errorbars in lattice plots.
It handles both x and y-errorbars as well as asymmetrical and symmetrical errorbars.
Sevarel features such as the whisker widths and other praghical parameters can
be controlled. Furthermore, the error bars can be offset from the main data 
point, and the inner part of the bars can be set not to be drawn (such that 
the main data point stands out). The function can also plot a single point 
without use of panel.xyplot or panel.superpose.

Provide either x.minus or x.plus or use x.err if the errorbar is 
symmetrical. Likewise for the y-errorbar. Apply NA values for errorbars
that should not be plotted (e.g. if they are within symbols).
   
Improvements: Now works with lattice dotplots.
See also \link{panel.binned.errorbars}, \link{panel.ebars} and \link{panel.ebars.grouped}.
}
\author{
Claus E. Andersen
}
