% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepanel-robust-functions.R
\name{prepanel.robust.y}
\alias{prepanel.robust.y}
\title{Robust prepanel function for lattice plots (ylim).}
\usage{
none
}
\arguments{
\item{x}{gives the x-values.}

\item{y}{gives the y-values.}

\item{...}{for additional parameters.}
}
\value{
a list with ylim-values for a trellis plot.
}
\description{
Lattice has a problem if there is only one single value and if
that value is very large (for example, y=6e66). The following
prepanel functions can be used to handle such situations.
See also \link(prepanel.robust.x) and \link(prepanel.robust.xy)
}
\author{
Claus E. Andersen
}
