% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-trellis-plots-function.R
\name{print.trellis.plots}
\alias{print.trellis.plots}
\title{Help place multiple lattice plots on a single page}
\usage{
plt0 <- xyplot(rnorm(10)~rnorm(10))
  print.trellis.plots(list(plt0,plt0,plt0),3,1)
}
\arguments{
\item{plt.list}{is the list of trellis plots. NULL is ok, also for single elements in the list.}

\item{cols}{specify a given number of columns}

\item{rows}{specify a given number of rows}

\item{x.size}{is the x-region to be used (0-1)}

\item{y.size}{is he y-region to be used (0-1)}

\item{x.overlap}{is the amount of x-overlap}

\item{y.overlap}{is the amount of y-overlap}

\item{plt.void}{is the default void plot may be supplied}

\item{use.plt.void:}{Control if some plot should be made where elements in the plt.list are NULL}

\item{trace:}{If true, the function with provide text output during execution}

\item{more:}{If false, the last plot in this call is the final one (see print.trellis). Otherwise
it is assumed that more stuff will be added to the plot.}
}
\value{
plt A Lattice plot showing the colors
}
\description{
Help place multiple lattice plots on a single page. Note that the
plots to be placed should be given as a list.
}
\author{
Claus E. Andersen
}
