% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder-for-trellis-function.R
\name{reorder.for.trellis}
\alias{reorder.for.trellis}
\title{Reorder a factor such that lattice plots come out in the correct numerical order.}
\usage{
reorder.for.trellis(c("OSL01 - 5 s", "OSL02 - 200", "OSL03 - -100.3", "A03 - -100.3 s"))
}
\arguments{
\item{x}{is the vector to be reordered}

\item{method}{is the methiod to be used. Only "Nth-number" has been implemented}

\item{absolute}{use absolute values. This avoids problems for text like temp-25-degC where
25 is probably 25 degC rather than -25 degC.}

\item{N=1}{means the first number (ignoring text etc.) N=2 means the second number (ignoring text etc.)}

\item{reverse:}{If true the order is reversed.}
}
\value{
A reordered vector
}
\description{
Reorder a factor such that lattice plots come out in the correct numerical order. This
function is useful when lattice plots are conditioned using factors with a mixture of text and numbers.
}
\author{
Claus E. Andersen
}

