% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-trellis-function.R
\name{set.trellis}
\alias{set.trellis}
\title{Define defaults for lattice plotting}
\usage{
set.trellis(col=c(8,6,6,6),pch=c(1,16))
  xyplot(I(rnorm(100))~rnorm(100))#' @name set.trellis
}
\arguments{
\item{col}{is the vecor of colors}

\item{lty}{is the vector of line types}

\item{lwd}{is the vector of line widths}

\item{cex}{is the vector of symbol sizes}

\item{font}{is the font}

\item{what}{is "all" or "plot.line"}

\item{trace:}{If true the function becomes verbose.}
}
\value{
void. The side effect is that the trellis/lattice parameters have changed.
}
\description{
Define defaults for lattice plotting (cf. settrellis in the Hmisc library)
}
\author{
Claus E. Andersen
}

