% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show.trellis.colors}
\alias{show.trellis.colors}
\title{Show available colors in Lattice graphs}
\usage{
show.trellis.colors(one.panel=TRUE)
}
\arguments{
\item{one.panel}{indicated if everything should be in one single panel. This will give a lot
of overlap among points, but could still be useful.}
}
\value{
plt A Lattice plot showing the colors
}
\description{
Show the various colors
}
\author{
Claus E. Andersen
}

