% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show.trellis.symbols}
\alias{show.trellis.symbols}
\title{Show available colors and symbols for Lattice graphics}
\usage{
show.trellis.symbols(palette(),"default palette")
}
\arguments{
\item{col.pal}{is the pallette Such as:
palette(cm.colors(10))
palette(heat.colors(11))
palette(terrain.colors(12))
palette(topo.colors(12))
palette("default"))}

\item{txt}{is the title of the plot}

\item{pch}{is a vector of symbols to be used}

\item{cex}{is the relative size of the symbols}
}
\value{
plt A Lattice plot showing the colors
}
\description{
Show the various colors and symbols in Trellis based on integer index
as given in palette().
}
\author{
Claus E. Andersen
}

