% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual-plot-demo-function.R
\name{trellis.residual.plot.demo}
\alias{trellis.residual.plot.demo}
\title{Function for demonstrating how to make residual plots in lattice}
\usage{
plt <- trellis.residual.plot.demo()
print.trellis.plots(list(plt),y.size=c(0.2,0.8))
}
\value{
A Lattice plot
}
\description{
This function uses several features of the latticeExtra package
to make plots with one large panel for the main data and a much smaller plot
above for the residuals. The data may me split into different levels. In the
specific example there are three laboratories, and so the final plot
consists of 3 x 2 panels. The main plots have the same scale, and likewise for 
the residual panels.
}
\author{
Claus E. Andersen
}
