% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtplot-demo-function.R
\name{txtplot.demo}
\alias{txtplot.demo}
\title{This function provides demonstrations of the use of texplot}
\usage{
txtplot.demo()
}
\value{
Lattice plots.
}
\description{
Demonstrations of how txtplot can be used for presentationof
text and dataframe.
}
\author{
Claus E. Andersen
}
