% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtplot-function.R
\name{txtplot}
\alias{txtplot}
\title{Make a Lattice graph of text and data}
\usage{
txtplot("hello world")
}
\arguments{
\item{txt}{is the text or the dataframe to be printed}

\item{cex}{(printing formatting)}

\item{col}{(printing formatting)}

\item{font}{(printing formatting).}

\item{new:}{If true, txtplot opens a new graphsheet for graphical output}

\item{start.position}{is a list with the start (x,y) position and an optional lineno on the page. 
start.pos = list(0.5,0.5, 6) refers to the center of the page and that the current line number is 6. 
The line number is used for keeping track of when we need a new page. If start.pos==NULL, then we 
just use the defaults x.start.default and y.start.default.}

\item{maintain.start.pos:}{If false, then start each new line at x=x.start.default otherwise 
start each new line at the orginal x-start.pos}

\item{new.line:}{If true,a new line is added after txt. Alternatively, include a backslash-n in the text.}

\item{page.size:}{number of lines before page break.}

\item{line.size:}{number of characters on a given line before line break (with or without hyphen)}

\item{line.size.extra:}{to promote that lines are broken where there is white space we allow for
a more relaxed break of lines. txtplot tries to break lines when there are more characters than 
given by line.size. However, it does not break before it reached a blank space - in so far that 
blank space occurs before the number of character is in the line is larger than: line.size + line.size.extra.}

\item{head:}{not implemented}

\item{foot:}{not implemented}

\item{evaluate:}{treat txt as commands, and execute these (has never been used).}

\item{adjust.table:}{add leading blanks using create.latex.table with the pretty.func argument}

\item{table.lines:}{a vector with line numbers where horizontal lines should be made. This only
applies to when txt is a data.frame.  999 is the last line. table.lines = TRUE is equivalent to
table.lines =c(1,2,999) table.lines = F gives no table lines. Note: The table lines are drawn 
using the lines-command and is therefore restricted to the x-y drawing panel (not the full 0-1 
by 0-1 coordinate system). In comparison, the text is drawn using the mtext-command. Set 
trace=TRUE and a red box will show where table lines can be drawn. If table lines are requested}

\item{table.line.pos:}{Fine control of where table lines should be placed.}

\item{col.sign:}{the sign between coloumns (used by create.latex.table)}

\item{hyphen.sign:}{the sign used by txtplot when it breaks lines.}

\item{char.width:}{distance between the start of each character (the size of each character is set by cex).}

\item{line.height:}{height of each line}

\item{box:}{list(coordinates, col, lwd, lty) or logical. The coordinates should be provided
in the same order as usr (i.e. as: c(x1,x2,y1,y2)).}

\item{char.groupA/B/C:}{Some characters do not have the same reference point as others, and we need to move these
somewhat. Each char.groupX is a list with three elements: a vector of characters, a number that specify the 
movements in the x and y directions (x.right and y.lift, respectively).}

\item{trace:}{debugging information}

\item{x.start.default:}{The default start x-position}

\item{y.start.default:}{The default start y-position}

\item{vp:}{The viewport that will be created for new pages.}
}
\value{
The current position is returned. This can be input on subsequent calls as start.pos.
}
\description{
This function can print text or dataframes on a separate graphsheet or on 
an excisting graph. The main idea behind the function is to add experimental or 
analytical details (=text) on separate 'graphs' in between the real graphs. 
When everything is exported to pdf or PowerPoint we both have explanations and graphs 
in the same document. At the same time, we can easily read the text in the script. 
Hence this function should facilitate improved documentation. This is kind of a 
simple way of literate programming.
}
\author{
Claus E. Andersen
}
