% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clanLattice}
\alias{clanLattice}
\title{Version function for the clanLattice library}
\usage{
clanLattice()
}
\value{
A list of information about the version and functions within clanLattice.
}
\description{
Version function for the clanLattice library.
}
\author{
Claus E. Andersen
}

