% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-ebars-demo-function.R
\name{panel.ebars.demo}
\alias{panel.ebars.demo}
\title{Demonstration of how to use panel.ebars for drawing error bars in lattice plots}
\usage{
panel.ebars.demo()
}
\value{
Lattice plot examples with error bars
}
\description{
Demonstration of how to use panel.ebars for drawing error bars in lattice plots
}
\author{
Claus E. Andersen
}
