% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscript-groups-panel-function-demo-function.R
\name{panel.function.demo}
\alias{panel.function.demo}
\title{Demonstration of how to write a panel function (subscripts and groups)}
\usage{
require(grid)
 require(lattice)
 pp <- panel.function.demo()
 pp$plt3
}
\value{
A Lattice plot
}
\description{
This function demonstrate how to write a panel function that 
takes care of subscripts and groups. The main trick is to always use:

 panel = function(x, y, subscripts, groups,...){ }
 
subscripts are always available, but groups may not be, so to mitigate that
simply use:

              if(missing(groups))\{
              panel.xyplot(x, y, ...)\} else \{
              panel.superpose(x, y, subscripts,groups, ...)
              \}#if 
The use of \code{panel.number()}, \code{current.row}, and \code{current.column()} is also demonstrated.
}
\author{
Claus E. Andersen
}
