% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reorder.factor}
\alias{reorder.factor}
\title{Reorder a vector of type factor}
\usage{
show.trellis.symbols(palette(),"default palette")
}
\arguments{
\item{Factor}{is the vector of factors to be reordered}

\item{x}{is a vector of numerical values with the same length as Factor.}

\item{Function}{is a function used on the aggregared data in x. The reordering will be
based on these results.}

\item{...}{Any other values that could be supplied to Function}
}
\value{
A reordered factor
}
\description{
Reorder a vector of type factor in accordance with the
results of groupwise results (e.g. mean values) for the different
levels in the factor.
}
\author{
Claus E. Andersen
}

