% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left-aligned-main-label-demo-function.R,
%   R/residual-plot-demo-function.R
\name{trellis.residual.plot.demo}
\alias{trellis.residual.plot.demo}
\alias{left.aligned.main.label.demo}
\title{Demonstration how tocreate a lattice plot with a left-aligned main label}
\usage{
require(lattice)
require(latticeExtra)
require(dplyr)
require(grid)
plt <- left.aligned.main.label.demo()
print(plt)

require(lattice)
require(latticeExtra)
require(dplyr)
require(grid)
plt <- trellis.residual.plot.demo()
print.trellis.plots(list(plt),y.size=c(0.2,0.8))
}
\value{
A Lattice plot

A Lattice plot
}
\description{
This functions creates a left-aligned main label
using a textgrob as given below.
How did I get that idea? See Lattice book p.152 / sec. 9.2.1 + demo("labels",package="lattice")

This function uses several features of the latticeExtra package
to make plots with one large panel for the main data and a much smaller plot
above for the residuals. The data may me split into different levels. In the
specific example there are three laboratories, and so the final plot
consists of 3 x 2 panels. The main plots have the same scale, and likewise for 
the residual panels.
}
\author{
Claus E. Andersen

Claus E. Andersen
}
