% Generated by roxygen2 (4.0.1): do not edit by hand
\name{txtplot.demo}
\alias{txtplot.demo}
\title{This function provides demonstrations of the use of texplot}
\usage{
txtplot.demo()
}
\value{
Lattice plots.
}
\description{
Demonstrations of how txtplot can be used for presentationof
text and dataframe.
}
\author{
Claus E. Andersen
}

